no_consume_or_assign ref_element_addr should not have any destroys
This fixes an issue when doing move-checking on a read accessor,
where the field is only borrowed. After the MoveOnlyAddressChecker
ran on it, it'd inject a destroy that didn't get "claimed":

```
  %2 = ref_element_addr %0 : $ListOfFiles, #ListOfFiles.file
  %3 = mark_must_check [no_consume_or_assign] %2 : $*File
  %4 = begin_access [read] [dynamic] %3 : $*File
  %5 = load_borrow %4 : $*File
  yield %5 : $File, resume bb1, unwind bb2

bb1:
  end_borrow %5 : $File
  end_access %4 : $*File
  destroy_addr %2 : $*File // BAD
  %9 = tuple ()
  return %9 : $()
```

The approach of this fix is to recognize that at the point we're
injecting destroys, we would have emitted diagnostics and stopped
already if there were any consuming uses that we need to clean-up
after, since we're in `no_consume_or_assign` checking mode here
when just reading the field.