[borrow-operator] Add support for immutable lvalues.
Any var construct that is accessed via the borrow operator will produce an
immutable lvalue. We will use this to ensure that we know when modifying the AST
that such lvalues should not be accessed as an rvalue and also we will use it to
validate that such things are not assigned to or inouted.