Playground transform should log function and closure parameters
Add logging of function and closure parameters when the playground transform is enabled.

MOTIVATION
The goal of the optional "playground transform" step in Sema is to instrument the code by inserting calls to `__builtin_log()` and similar log functions, in order to record the execution of the compiled code.  Some IDEs (such as Xcode) enable this transform by passing `-playground` and provide implementations of the logger functions that record information that can then be shown in the IDE.

The playground transform currently logs variable assignments and return statements, but it doesn't log the input parameters received by functions and closures.  Knowing these values can be very useful in order to understand the behaviour of functions and closures.

CHANGES
- add a `ParameterList` parameter to `InstrumenterSupport::transformBraceStmt()`
   - this is an optional parameter list that, if given, specifies the parameters that should be logged at the start of the brace statement
   - this has to be passed into the function because it comes from the owner of the `BraceStmt`
- adjust _PlaygroundTransform.cpp_ to make use of this list
   - the transform will insert calls to `__builtin_log()` for each of the parameters, in order
- adjust _PCMacro.cpp_ to accept the parameter, though this instrumenter doesn't currently make use of the new information
- add two new unit tests (one for functions and one for closures)
- adjust four existing unit tests to account for the new log calls

QUESTIONS
- should this be off by default and instead be controllable by a new frontend flag?
   - this is the first of a series of log calls I plan to add to the playground transform, and so it might be useful to control them using some kind of organized option set

rdar://104974072