[NFC] Rework the internal design of SIL type lowering.
The old, CRTP-based design was really difficult to understand
because of how it tried to map type-based visitation into property-based
handling.  That didn't really work out for producing TypeLowering
objects anyway, since a lot of different type kinds want their own
TypeLowering handling even if they share abstract properties.
That in turn led to a lot of redundancy between the two users,
and that redundancy tended to lead to bugs --- I got into this
because I kept finding problems where the pack-related types had
different properties on different paths.

That's bad enough, really, but there were other problems, too.
Classification basically repeated the entire AbstractionPattern-based
walk over the type that lowering does, and I'm not sure why --- there
aren't actually any type properties that vary by abstraction pattern,
and I'm not sure why there ever would be.  And the way that property
computation was directly tied to producing a TypeLowering object
meant that preserving certain kinds of top-down property was
unnecessarily cumbersome, like remembering that we looked through
an opaque archetype in the is-expansion-sensitive bit.  We also
weren't set up to answer any of the basic lowering questions (type
properties, lowered type) without vending a TypeLowering object,
even though we mostly do not need TypeLowering objects in clients
(and probably ought to get rid of them completely someday).

(The old design was entirely my fault, just to be clear here.)

Anyway, we can instead separate these into three related actions:
- producing a lowered type
- computing the properties of a lowered or formal type
- creating a TypeLowering object

Each of these operations is then relatively straightforward to implement
and largely stands separate from the others.  And we can make queries
for lowered types and certain type properties answerable without getting
a TypeLowering, although I have not changed clients to stop asking for a
TypeLowering unnecessarily.

I did have an initial design for type property computation which relied
on being given a lowered type, but that was unnecessary, expensive, and
caused problems with the infinite-type check.  There's no reason we can't
answer questions about type properties with just a type; we don't even
need an abstraction pattern for this.  (We do need a generic signature,
if the type contains type parameters.)

I removed the previous thing where it would find generic signatures for
interface substTypes in the abstraction pattern.  This required updating
a few clients, but the result is *so* much better and more obviously
correct.

Oh, and I added move-only-ness as a type property, although I'm not
sure my logic is completely correct in the presence of stuff like
SILMoveOnlyWrappedType.