[Macros] Call `visitAuxiliaryDecls` on every decl in `ASTWalker`
Hoist `visitAuxiliaryDecls` from `ASTWalker::visitMacroExpansionDecl` to `ASTWalker::doIt` to make sure all auxiliary decls are covered.

Because not all auxiliary decls are macro expansions, we need to add a `includeMacroExpansions` parameter to `visitAuxiliaryDecls` to disable macro expansions when the walker is asked not to walk macro expansions. This cannot be done via checking in the closure body of `visitAuxiliaryDecls`, because expanding a macro needs parent contexts' local discriminators, which do not exist when it's inside of a `#if false`; in order words, we need to prevent expanding a macro instead of skipping a visit.