Teach RValue to not recursively expand tuples with pack expansions
Arguably, it would be more consistent with the current design to
figure out a representation for pack-expansion components so that
we can keep them around in whatever storage.  But honestly, I
don't like the current design; I think the eager explosion is an
over-complicated mistake.  So this is partly just me finding an
excuse to not extend that to more cases.