Automatically update the CXX_STANDARD
Folks are running into issues updating because the CXX standard is
cached and we updated it. This patch allows CMake to update it to a
minimum required version for Swift. If it's in the cache and too low, it
will update it. If it's manually specified and too low, it will give a
fatal error letting the dev know it's too low.