Fix the minimal/lto configuration
We need to disable opaque pointers when we compiler the runtime the
linker complains:

```
  ld: Opaque pointers are only supported in -opaque-pointers mode
```

Alternatively, we could pass the opaque-pointers flag to the linker but
then it would complain about the swiftc generated files which still use
typed pointers.
Until swiftc is fixed, disable opaque pointers for .cpp runtime files.

rdar://106515243