Add --enable-asan option in swift-driver
This is to support building the swift-driver with -sanitize=address when the Swift build preset specifies enable-asan. Without this option and with enable-asan in picture the entire Swift build fails due to ASAN initialization failure in swift-driver.
This commit is dependent on https://github.com/apple/swift-driver/pull/1310 being merged first

rdar://104661463