Tie Swift backtracing pieces together in build
This patch ties the runtime and stdlib backtracing pieces together.
The `__swift_isThunkFunction` SPI is provided in the backtracing
runtime, which is gated behind `SWIFT_ENABLE_BACKTRACING`. If that flag
is turned off, the `Backtracing` part of the stdlib fails to build
because the symbols are missing, causing a build failure.

I've also added the explicit `option(SWIFT_ENABLE_BACKTRACING` option so
that incremental re-runs of CMake get the flag added with a nice
comment and a boolean value. The option is defaulted to OFF, though
build-script is defaulting it to ON.